+++ Makefile
@@ -1,7 +1,7 @@
 NAME	= minishell
 
 SRCS	= $(addprefix srcs/,$(addsuffix .c, $(FILES)))
-FILES	= pipex/pipex pipex/build_command_list pipex/utils_free_close \
+FILES	= pipex/pipex pipex/build_command_list pipex/utils_free_close pipex/utils_pipe \
 	pipex/utils_misc pipex/here_doc pipex/free_command_list pipex/check_builtin \
 	built_ins/ft_cd built_ins/ft_pwd built_ins/ft_env built_ins/ft_echo \
 	built_ins/ft_exit built_ins/ft_export built_ins/ft_unset built_ins/env_helper built_ins/env_helper_2\
