+++ srcs/pipex/free_command_list.c
@@ -6,7 +6,7 @@
 /*   By: yioffe <yioffe@student.42lisboa.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/24 13:09:17 by yioffe            #+#    #+#             */
-/*   Updated: 2024/05/22 08:28:21 by yioffe           ###   ########.fr       */
+/*   Updated: 2024/05/31 12:45:38 by yioffe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,17 +30,13 @@ void	free_command_list(t_arg **command_list)
 		}
 		free(current->arguments);
 		current->arguments = NULL;
-		//free(current->args);
-		//current->old_args_old = NULL;
 		free(current->path);
 		current->path = NULL;
 		next = current->next;
-		//free(current);
 		current = NULL;
 		current = next;
 	}
 	*command_list = NULL;
-	//close_all_protected();
 }
 
 void	dup_close(int fd, int reference)
@@ -48,9 +44,7 @@ void	dup_close(int fd, int reference)
 	if (fd != reference)
 	{
 		if (dup2(fd, reference) == -1)
-            perror("dup_close failure");
-		//printf("closing fd: %d\n", fd);
+			perror("dup_close failure");
 		ft_close(fd);
 	}
 }
-
