+++ srcs/pipex/utils_free_close.c
@@ -6,7 +6,7 @@
 /*   By: yioffe <yioffe@student.42lisboa.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/09 17:02:16 by yioffe            #+#    #+#             */
-/*   Updated: 2024/05/28 12:35:38 by yioffe           ###   ########.fr       */
+/*   Updated: 2024/05/31 12:34:47 by yioffe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,7 @@ void	ft_close(int fd)
 {
 	if (fd != -1 && fd != STDERR_FILENO && fd != STDIN_FILENO
 		&& fd != STDOUT_FILENO)
-	{
-		// printf("closing fd: %d\n", fd);
 		close(fd);
-	}
 }
 
 void	close_all_protected(t_shell *shell)
@@ -27,7 +24,6 @@ void	close_all_protected(t_shell *shell)
 	int	i;
 
 	i = 3;
-	// printf("closing all protected\n");
 	while (i < 1024)
 	{
 		if (i != shell->std_fds[0] && i != shell->std_fds[1]
@@ -42,20 +38,11 @@ void	close_all_unprotected(void)
 	int	i;
 
 	i = 0;
-	// printf("closing all not protected\n");
 	while (i < 1024)
 	{
 		close(i);
 		i++;
 	}
-}
-
-void	close_both_ends(int fd[2], bool pipe_error)
-{
-	if (pipe_error)
-		perror("Error creating pipe");
-	ft_close(fd[FD_IN]);
-	ft_close(fd[FD_OUT]);
 }
 
 void	exit_pipe_error(int fd[2])
@@ -64,5 +51,19 @@ void	exit_pipe_error(int fd[2])
 	{
 		perror("Error creating pipe");
 		exit(EXIT_FAILURE);
+	}
+}
+
+void	close_pipes(int *fd_pipe)
+{
+	if (fd_pipe[FD_IN] >= 0)
+	{
+		close(fd_pipe[FD_IN]);
+		fd_pipe[FD_IN] = -1;
+	}
+	if (fd_pipe[FD_OUT] >= 0)
+	{
+		close(fd_pipe[FD_OUT]);
+		fd_pipe[FD_OUT] = -1;
 	}
 }
